requirements.txt
google-genai
dotenv
numpy
langgraph
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
"unstructured[md]"
nltk
sentence-transformers
fastapi
uvicorn
python-telegram-bot